Navigational elements like anchors and button were still tab-able in the modal when it was closed (hidden). This also includes the modal close button. The tabindex="-1" on the modal itself gets ignored by Chrome and Firefox.
The visibility is now hidden by default on the modal making all children of the modal non tab-able. The visibility instantly gets set to visible when the modal opens and set to hidden again when the modal closes. This happens with a transition-delay identical to the transition-duration of the css opacity fadeout to preserve the transition.